Split into separate "ui" and "audio" crates.
This means that the top level of this project is a cargo workspace, instead
of a crate. Some minimal moving of code was done to allow this to compile.

The objective here is for dsp code to be in a separate lib crate that can
be used in Jupyter notebooks.